+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Craated Category List
 - Created veicule Routes
 - Added Multer
+- configured banner upload
+- configured cadaster of veicules
 
 ### Fixed